Set reassoc response flag based on the current state
While processing the assoc event, check the current wpa_supplicant state
and if supplicant is already in the WPA_COMPLETED state, set the
reassociation response flag in wpa_ft_parse_ies() to true. This is for
parsing the IE correctly in case of MLO.

Bug: 415472691
Test: Manual - Tested FT roaming case with MLO supported AP and confirmed
that the supplicant is sending the state change events to framework
while roaming.

Change-Id: Ib5402d60b7817a3b434f54bd52e5fba18637d1c4